+++ draft-packages/pagination/KaizenDraft/Pagination.tsx
@@ -1,21 +1,23 @@
-import React, { RefObject, useCallback } from "react"
+import React from "react"
 import arrowBackward from "@kaizen/component-library/icons/arrow-backward.icon.svg"
 import arrowForward from "@kaizen/component-library/icons/arrow-forward.icon.svg"
 import { Icon } from "@kaizen/component-library"
 import cx from "classnames"
-// We don't use this here
-import { Link } from "react-router"
-// import { scrollUpToElementIntoView } from "../../utils/dom"
-// import { defaultScrollOffset } from "../../constants/scrolling"
 import PageIndicator from "./PageIndicator"
 import TruncateIndicator from "./TruncateIndicator"
 import styles from "./styles.scss"
-import { createRange, getPaginationUrl } from "./utils"
+
+export const createRange = (start: number, end: number): number[] => {
+  const range: number[] = []
+  for (let i = start; i <= end; i++) {
+    range.push(i)
+  }
+  return range
+}
 
 type Props = {
   currentPage: number
   pageCount: number
-  // location: string
   ariaLabelNextPage?: string
   ariaLabelPreviousPage?: string
   ariaLabelPage?: string
@@ -25,8 +27,6 @@ type Props = {
   // contentRef: RefObject<HTMLElement>
   onPageChange: (newPage: number) => void
 }
-
-// TODO: this can be part of the props.
 
 const boundaryPagesRange = 1
 const siblingPagesRange = 1
@@ -38,19 +38,8 @@ const Pagination = ({
   ariaLabelPreviousPage,
   ariaLabelPage,
   onPageChange,
-}: // location,
-// contentRef,
-Props) => {
-  // const getPageHref = (page: number) => getPaginationUrl(location, page)
-
+}: Props) => {
   // Click event for all pagination buttons (next, prev, and the actual numbers)
-  // const handleButtonClick = useCallback(() => {
-  //   // const content = contentRef.current
-  //   // if (!content) return // hopefully this shouldn't happen
-  //   // scrollUpToElementIntoView(content, { offset: defaultScrollOffset })
-  //   // }, [contentRef])
-  // }, [])
-
   const handleButtonClick = (newPage: number | "prev" | "next") => {
     if (newPage === "prev") {
       onPageChange(currentPage - 1)
@@ -67,7 +56,6 @@ Props) => {
     <PageIndicator
       page={page}
       selected={currentPage === page}
-      // location={location}
       ariaLabelPage={ariaLabelPage}
       onPageClick={handleButtonClick}
     />
@@ -155,7 +143,6 @@ Props) => {
         className={cx(styles.arrowIconWrapper, {
           [styles.arrowIconWrapperDisabled]: previousPageDisabled,
         })}
-        // to={getPageHref(previousPageDisabled ? currentPage : currentPage - 1)}
         aria-label={ariaLabelPreviousPage}
         aria-disabled={previousPageDisabled}
         onClick={() => handleButtonClick("prev")}
@@ -169,7 +156,6 @@ Props) => {
         className={cx(styles.arrowIconWrapper, {
           [styles.arrowIconWrapperDisabled]: nextPageDisabled,
         })}
-        // to={getPageHref(nextPageDisabled ? currentPage : currentPage + 1)}
         aria-label={ariaLabelNextPage}
         aria-disabled={nextPageDisabled}
         onClick={() => handleButtonClick("next")}
