+++ draft-packages/pagination/package.json
@@ -34,10 +34,7 @@
   "dependencies": {
     "@kaizen/component-library": "^10.0.3",
     "classnames": "^2.3.1",
-    "lodash": "^4.17.21",
-    "path-to-regexp": "^6.2.0",
-    "query-string": "^7.0.1",
-    "react-router": "^5.2.0"
+    "path-to-regexp": "^6.2.0"
   },
   "devDependencies": {},
   "peerDependencies": {
